getrepos refactored
