+++ src/context/github/githubReducer.js
@@ -26,6 +26,13 @@ export default (state, action) => {
         users: [],
         loading: false
       };
+    case GET_REPOS: {
+      return {
+        ...state,
+        repos: action.payload,
+        loading: false
+      };
+    }
     case SET_LOADING:
       return {
         ...state,
